Fix for Export-Excel #165
Fixed handling of the TableName with parameter validation, we now throw
an error in case the TableName:
- starts with something else then a letter
- is NULL or empty
- Contains spaces
Just like in the orriginal EPPlus.dll library.

Added Try/Catch clause for the End clause. So errors like 'TableName is
not unique' are terminating errors and the function isn't exectued any
further.

Also cleaned up for readability some cosmetic stuff:
- 'if(' to 'if ('
- double quotes to single quotes where needed