Update Get-HTMLTable
Added logic to check whether the file is local (UNC Path or Local Drive) or Web Hosted (via regex). Minor changes to the way the for loop works to account for the new objects and to make sure they are both compatible with either local or web hosted html tables.