Some more changes
Added a regex param validator for the $Path parameter: It must have 'xlsx' in the path or the function will not work, the Export-Excel function requires the file be in xlsx format. 

Added some logic around if the path is set, and Append is not so that the user is aware they may be overwriting an already existing file and if they would like to proceed. Yes will overwrite the file, No will break.

$TempFile variable is added to account for the Append state, if a temp file is used append will have no affect anyway, so is set to $False. If $TempFile is set to $True and Append is used it will set the $AppendState = $True.

If $Show is not set and $TempFile is set to $False the path is sent to the clipboard and written to the console (via Write-Host).

Export-Excel is at the end of the Get-HTMLTable Pipeline instead of setting a variable