Update to parse date strings as OLE Date
Ran into some issues while importing a fairly large Excel document and the date time formatting / column definitions within it not being consistent, this is the fix that worked to check and convert strings that normally "[DateTime]::FromOADate()" couldn't parse because well they're strings and not OLE Dates. It seems like a good addition to beef up the date parsing from imported documents to lessen the burden of defining column types in Excel but that is only my own opinion.